// PROJECT WEB2/Angular/src/app/shared/models/registration.model.ts
export class Registration{
  username:string = "";
  email: string = "";
  password: string = "";
  nameAndLastName: string = "";
  dateOfBirth: string = "";
  address: string = "";
  type:  string= "";
  password2: string = "";
  validationDeliverers: boolean= false;
  idDeliverer: boolean=false;
  image:string="";


}